[feat] Create Contest Card component and finish contest-select page
